+++ types/index.d.ts
@@ -1,0 +1,3 @@
+import DLL from './lib/dll';
+import { AccessRestrictedDLLItem } from './lib/dllItemAccessRestrictor';
+export { DLL, AccessRestrictedDLLItem as DLLItem, };