+++ types/src/index.d.ts
@@ -1,3 +1,0 @@
-import DLL from './lib/dll';
-import DLLItem from './lib/dllItem';
-export { DLL, DLLItem, };